feat: Add initialize function to set up config state
This avoids having to provide the reused state on every call to the search API and can be reused
among other methods.

The state management here revolves around using ESM modules, which have private scope, and are run
*once* when first imported. Its similar to how a dependency injection with a singleton class would
work, the module logic executes once to initiate state and then one can use the provided module API
to interact with that state.