fix!: Search API functions how take query as a parameter
The query to call the APIs with is a very dynamic (as in it changes a lot) value, as opposed to most
of the other options which are more static. This change allows callers to call the function with the
query as string, separate from the other options. It also allows for cleaner typing in consuming
libraries with `ProductSearchOptions` as opposed to `Omit<ProductSearchOptions, 'q'>` when the query
comes form some usually user generated state.

BREAKING CHANGE: `productSearch`, `bestseller`, `contentSearch`, `categorySearch` and `autoSuggest`
parameters have changed.